Factor out utility functions into `util` module.
