+++ ethcore/src/engines/authority_round/mod.rs
@@ -49,6 +49,7 @@ use unexpected::{Mismatch, OutOfBounds};
 
 mod finality;
 mod randomness;
+mod util;
 
 /// `AuthorityRound` params.
 pub struct AuthorityRoundParams {
