+++ ethcore/src/engines/authority_round/randomness.rs
@@ -2,11 +2,10 @@
 //!
 //! This module contains the support code for the on-chain randomness generation used by AuRa.
 
-use std::fmt;
-
 use ethereum_types::Address;
 
-use client::{BlockId, CallContract, Client};
+use super::util::{BoundContract, CallError};
+use client::{BlockId, Client};
 
 use_contract!(aura_random, "res/authority_round_random.json");
 
@@ -52,7 +51,6 @@ pub enum RandomnessPhase {
 	Reveal,
 }
 
-
 /// Phase loading error for randomness generation state machine.
 ///
 /// This error usually indicates a bug in either the smart contract or the phase loading function.
@@ -64,66 +62,7 @@ pub enum PhaseError {
 	/// commit phase.
 	RevealedInCommit,
 	/// Calling a contract to determine the phase failed.
-	CallError(String),
-	/// Decoding a called contract's return value failed while determinig the phase.
-	DecodeError(ethabi::Error),
-}
-
-
-/// Utility structure for binding contracts.
-///
-/// A bound contract is a combination of a `Client` reference, a `BlockId` and a contract `Address`.
-/// These three parts are enough to call a contract's function; return values are automatically
-/// decoded.
-pub struct BoundContract<'a> {
-	client: &'a Client,
-	block_id: BlockId,
-	contract_addr: Address,
-}
-
-impl<'a> fmt::Debug for BoundContract<'a> {
-	fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
-		fmt.debug_struct("BoundContract")
-			.field("client", &(self.client as *const Client))
-			.field("block_id", &self.block_id)
-			.field("contract_addr", &self.contract_addr)
-			.finish()
-	}
-}
-
-impl<'a> BoundContract<'a> {
-	/// Create a new `BoundContract`.
-	#[inline]
-	pub fn bind(client: &Client, block_id: BlockId, contract_addr: Address) -> BoundContract {
-		BoundContract {
-			client,
-			block_id,
-			contract_addr,
-		}
-	}
-
-	/// Perform a function call to an ethereum machine.
-	///
-	/// Runs a constant function call on `client`. The `call` value can be serialized by calling any
-	/// api function generated by the `use_contract!` macro.
-	pub fn call_const<D>(&self, call: (ethabi::Bytes, D)) -> Result<D::Output, PhaseError>
-	where
-		D: ethabi::FunctionOutputDecoder,
-	{
-		let (data, output_decoder) = call;
-
-		let call_return = self
-			.client
-			.call_contract(self.block_id, self.contract_addr, data)
-			.map_err(PhaseError::CallError)?;
-
-		// Decode the result and return it.
-		// NOTE: We return an already wrapped error here, ideally this function could be generalized by
-		//       removing the `map_err` and sticking it elsewhere.
-		output_decoder
-			.decode(call_return.as_slice())
-			.map_err(PhaseError::DecodeError)
-	}
+	LoadFailed(CallError),
 }
 
 impl RandomnessPhase {
@@ -132,7 +71,7 @@ impl RandomnessPhase {
 	/// Calls various constant contract functions to determine the precise state that needs to be
 	/// handled (that is, the phase and whether or not the current validator still needs to send
 	/// commitments or reveal secrets).
-	fn load(
+	pub fn load(
 		client: &Client,
 		block_id: BlockId,
 		contract_address: Address,
@@ -141,21 +80,29 @@ impl RandomnessPhase {
 		let contract = BoundContract::bind(client, block_id, contract_address);
 
 		// Determine the current round and which phase we are in.
-		let round = contract.call_const(aura_random::functions::current_collect_round::call())?;
-		let is_reveal_phase =
-			contract.call_const(aura_random::functions::is_reveal_phase::call())?;
-		let is_commit_phase =
-			contract.call_const(aura_random::functions::is_commit_phase::call())?;
+		let round = contract
+			.call_const(aura_random::functions::current_collect_round::call())
+			.map_err(PhaseError::LoadFailed)?;
+		let is_reveal_phase = contract
+			.call_const(aura_random::functions::is_reveal_phase::call())
+			.map_err(PhaseError::LoadFailed)?;
+		let is_commit_phase = contract
+			.call_const(aura_random::functions::is_commit_phase::call())
+			.map_err(PhaseError::LoadFailed)?;
 
 		// Ensure we are not committing or revealing twice.
-		let committed = contract.call_const(aura_random::functions::is_committed::call(
-			round,
-			our_address,
-		))?;
-		let revealed: bool = contract.call_const(aura_random::functions::sent_reveal::call(
-			round,
-			our_address,
-		))?;
+		let committed = contract
+			.call_const(aura_random::functions::is_committed::call(
+				round,
+				our_address,
+			))
+			.map_err(PhaseError::LoadFailed)?;
+		let revealed: bool = contract
+			.call_const(aura_random::functions::sent_reveal::call(
+				round,
+				our_address,
+			))
+			.map_err(PhaseError::LoadFailed)?;
 
 		// With all the information known, we can determine the actual state we are in.
 		if is_reveal_phase && is_commit_phase {
